requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
click==6.6
ecdsa==0.13               # via python-jose
elasticsearch-dsl==5.0.0
elasticsearch==5.0.1
enum34==1.1.6
flask-sqlalchemy==2.1
flask==0.11.1
functools32==3.2.3.post2  # via jsonschema
future==0.16.0            # via python-jose
geopy==1.11.0
google-api-python-client==1.5.5
gunicorn==19.6.0
httplib2==0.9.2           # via google-api-python-client, oauth2client
itsdangerous==0.24        # via flask
jinja2==2.8               # via flask
jsonschema==2.5.1
markupsafe==0.23          # via jinja2
oauth2client==3.0.0
oauth2client==2.1.0
ordered-set==2.0.1
pathlib2==2.1.0
psycopg2==2.6.2
pyasn1-modules==0.0.8     # via oauth2client
pyasn1==0.1.9             # via oauth2client, pyasn1-modules, rsa
pycrypto==2.6.1           # via python-jose
python-dateutil==2.6.0    # via elasticsearch-dsl
python-decouple==3.0
python-dotenv==0.6.1
python-jose==1.3.2
python-magic==0.4.12
pytz==2016.7
PyYAML==3.12
rsa==3.4.2                # via oauth2client
six==1.10.0               # via elasticsearch-dsl, google-api-python-client, oauth2client, pathlib2, python-dateutil, python-jose, sqlalchemy-utils, swagger-spec-validator
SQLAlchemy-Utils==0.32.11
SQLAlchemy==1.1.4         # via flask-sqlalchemy, sqlalchemy-utils
swagger-spec-validator==2.0.2
typing==3.5.2.2
unicodecsv==0.14.1
uritemplate==3.0.0        # via google-api-python-client
urllib3==1.19.1           # via elasticsearch
Werkzeug==0.11.11         # via flask
wrapt==1.10.8

# The following packages are considered to be unsafe in a requirements file:
# setuptools                # via swagger-spec-validator
